cast rendererID to uint32_t for opengl

+++ Engine/src/Platform/OpenGL/OpenGLFrameBuffer.cpp
@@ -130,14 +130,14 @@ void OpenGLFrameBuffer::Generate()
 
 			m_ColourAttachments[i] = Texture2D::Create(m_Specification.width, m_Specification.height, format, m_Specification.samples);
 
-			BindTexture(multisample, m_ColourAttachments[i]->GetRendererID());
+			BindTexture(multisample, (uint32_t)m_ColourAttachments[i]->GetRendererID());
 			switch (m_ColourAttachmentSpecifications[i].textureFormat)
 			{
 			case FrameBufferTextureFormat::RGBA8:
-				AttachColourTexture(m_ColourAttachments[i]->GetRendererID(), m_Specification.samples, GL_RGBA8, GL_RGBA, m_Specification.width, m_Specification.height, i);
+				AttachColourTexture((uint32_t)m_ColourAttachments[i]->GetRendererID(), m_Specification.samples, GL_RGBA8, GL_RGBA, m_Specification.width, m_Specification.height, i);
 				break;
 			case FrameBufferTextureFormat::RED_INTEGER:
-				AttachColourTexture(m_ColourAttachments[i]->GetRendererID(), m_Specification.samples, GL_R32I, GL_RED_INTEGER, m_Specification.width, m_Specification.height, i);
+				AttachColourTexture((uint32_t)m_ColourAttachments[i]->GetRendererID(), m_Specification.samples, GL_R32I, GL_RED_INTEGER, m_Specification.width, m_Specification.height, i);
 				break;
 			default:
 				break;
@@ -153,11 +153,11 @@ void OpenGLFrameBuffer::Generate()
 		auto format = FrameBufferFormatToTextureFormat(m_DepthAttachmentSpecification.textureFormat);
 		m_DepthAttachment = Texture2D::Create(m_Specification.width, m_Specification.height, format, m_Specification.samples);
 	
-		BindTexture(multisample, m_DepthAttachment->GetRendererID());
+		BindTexture(multisample, (uint32_t)m_DepthAttachment->GetRendererID());
 		switch (m_DepthAttachmentSpecification.textureFormat)
 		{
 		case FrameBufferTextureFormat::DEPTH24STENCIL8:
-			AttachDepthTexture(m_DepthAttachment->GetRendererID(), m_Specification.samples, GL_DEPTH24_STENCIL8, GL_DEPTH_STENCIL_ATTACHMENT, m_Specification.width, m_Specification.height);
+			AttachDepthTexture((uint32_t)m_DepthAttachment->GetRendererID(), m_Specification.samples, GL_DEPTH24_STENCIL8, GL_DEPTH_STENCIL_ATTACHMENT, m_Specification.width, m_Specification.height);
 			break;
 		default:
 			break;
@@ -259,6 +259,5 @@ void OpenGLFrameBuffer::ClearAttachment(size_t index, int value)
 
 	GLenum textureFormat = TextureFormatToOpenGlTextureFormat(spec.textureFormat);
 
-	glClearTexImage(m_ColourAttachments[index]->GetRendererID(), 0, textureFormat, GL_INT, &value);
+	glClearTexImage((uint32_t)m_ColourAttachments[index]->GetRendererID(), 0, textureFormat, GL_INT, &value);
 }
-
